refactor(finance): complete BigDecimal migration across the project

+++ backend/marketfusion/src/main/java/com/marketfusion/controller/AnalyticsController.java
@@ -4,11 +4,9 @@ import com.marketfusion.service.AnalyticsService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
@@ -22,13 +20,13 @@ public class AnalyticsController {
 
     @Operation(summary = "Выручка за последние 30 дней")
     @GetMapping("/revenue")
-    public ResponseEntity<Double> getRevenueLast30Days() {
+    public ResponseEntity<BigDecimal> getRevenueLast30Days() {
         return ResponseEntity.ok(analyticsService.getRevenueLast30Days());
     }
 
     @Operation(summary = "Выручка по дням за последние 30 дней")
     @GetMapping("/daily-revenue")
-    public ResponseEntity<Map<LocalDateTime, Double>> getDailyRevenueLast30Days() {
+    public ResponseEntity<Map<LocalDateTime, BigDecimal>> getDailyRevenueLast30Days() {
         return ResponseEntity.ok(analyticsService.getDailyRevenueLast30Days());
     }
